requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
tensorflow
pillow==10.1.0
numpy==1.26.0
python-dotenv==1.0.0
gunicorn==21.2.0
